Made the tests work again Changed from float to int. (Bc of database) Added Question Points in exam DAO Added PluginVersion in questionDAO

+++ src/main/kotlin/nl.han.ica/examplatform/models/question/Question.kt
@@ -32,7 +32,7 @@ data class Question(
         @ApiModelProperty(notes = "Text of the question. This could be null if the question has subquestions")
         val questionText: String? = null,
         @ApiModelProperty(notes = "Points assigned to a question in an Exam")
-        val questionPoints: Float? = null,
+        val questionPoints: Int? = null,
         @ApiModelProperty(notes = "ID of the course")
         val courseId: Int? = null,
         @ApiModelProperty(notes = "Type of exam the question is for.")
